requirements.txt
numpy==1.19.1
pandas==1.4.2
opencv-python==4.5.5.64
tensorflow==2.6.2
pydantic==1.9.1
paddlepaddle==2.3.0
paddleocr==2.5.0.3
Pillow==9.1.1
html2text==2020.1.16
# cairosvg
fastapi==0.70.1
tensorflow-hub==0.12.0
uvicorn==0.16.0
scipy==1.8.1
PyMySQL==1.0.2
multimethod==1.8
kafka-python==2.0.2
APScheduler==3.9.1
python-multipart-0.0.5
